Further rework file structure and resolve analysis errors
